requirements.txt
pandas
smop
dash==1.11.0
dash_html_components==1.0.3
mysql_connector_repackaged==0.3.1
gunicorn
